% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineForecast.R
\name{DefineForecast}
\alias{DefineForecast}
\title{Define Fertility Forecast}
\usage{
DefineForecast(
  jumpoff_asfrs,
  forecast_horizon,
  ages,
  wlast,
  target_tfr,
  target_mab,
  asfr_growth_rate,
  timestep_of_steepest_growth,
  randomness = c("finland1995-2024", "empirical", "manual")
)
}
\arguments{
\item{jumpoff_asfrs}{numeric vector of ASFRs in jumpoff year}

\item{forecast_horizon}{integer scalar defining the number of
timesteps to forecast}

\item{ages}{numeric vector of starting ages for age groups}

\item{wlast}{width of the last age group}

\item{target_tfr}{Target TFR}

\item{target_mab}{Target mean age at birth}

\item{asfr_growth_rate}{Logistic growth rate of ASFRs over
forecasting horizon}

\item{timestep_of_steepest_growth}{Timesteps into forecast until
half of ASFR change has been realized}

\item{randomness}{Specification of randomness around future ASFRs.
One of 'finland1995-2024', 'empirical', 'manual'. See details.}
}
\value{
A PFS object.
}
\description{
Define Fertility Forecast
}
\examples{
DefineForecast(
    jumpoff_asfrs = c(0.0031, 0.0269, 0.0671, 0.0888, 0.0512, 0.0134, 0.0012),
    forecast_horizon = 30,
    ages = c(15, 20, 25, 30, 35, 40, 45),
    wlast = 5,
    target_tfr = 1.51,
    target_mab = 32.9,
    asfr_growth_rate = 0.3,
    timestep_of_steepest_growth = 15,
    randomness = 'finland1995-2024'
 )
}
