% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeForecast.R
\name{MakeForecast}
\alias{MakeForecast}
\title{Make ASFR/TFR Forecasts}
\usage{
MakeForecast(forecast_definition, nsim = 250)
}
\arguments{
\item{forecast_definition}{A pfs object.}

\item{nsim}{An integer scalar for the number of simulated ASFR/TFR
paths.}
}
\value{
A list with 7 entries
}
\description{
Make ASFR/TFR Forecasts
}
\examples{
forecast_definition <- DefineForecast(
    jumpoff_asfrs = c(0.0031, 0.0269, 0.0671, 0.0888, 0.0512, 0.0134, 0.0012),
    forecast_horizon = 30,
    ages = c(15, 20, 25, 30, 35, 40, 45),
    wlast = 5,
    target_tfr = 1.51,
    target_mab = 32.9,
    asfr_growth_rate = 0.3,
    timestep_of_steepest_growth = 15,
    randomness = 'finland1995-2024'
 )
 MakeForecast(forecast_definition)
}
