% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictTargetASFRs.R
\name{PredictTargetASFRs}
\alias{PredictTargetASFRs}
\title{Predict Target ASFRs}
\usage{
PredictTargetASFRs(
  forecast_definition,
  target_ratio = c(0.697, 0.842, 0.98, 1.499, 1.888, 1.489, 1.917),
  weights = c(tfr = 10, mab = 10, ratios = 1)
)
}
\arguments{
\item{forecast_definition}{A pfs object.}

\item{target_ratio}{.}

\item{weights}{A numeric vector of positive weights stating the
relative importance of the TFR, MAB, and ratios optimization
criteria.}
}
\value{
A list of target ASFRs and further details regarding the
optimization.
}
\description{
Predict Target ASFRs
}
\examples{
forecast_definition <- DefineForecast(
    jumpoff_asfrs = c(0.0031, 0.0269, 0.0671, 0.0888, 0.0512, 0.0134, 0.0012),
    forecast_horizon = 30,
    ages = c(15, 20, 25, 30, 35, 40, 45),
    wlast = 5,
    target_tfr = 1.51,
    target_mab = 32.9,
    asfr_growth_rate = 0.3,
    timestep_of_steepest_growth = 15,
    randomness = 'finland1995-2024'
 )
PredictTargetASFRs(forecast_definition)
}
