% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictTrajectoryToTargetASFRs.R
\name{PredictTrajectoryToTargetASFRs}
\alias{PredictTrajectoryToTargetASFRs}
\title{Predict Trajectory to Target ASFRs}
\usage{
PredictTrajectoryToTargetASFRs(
  forecast_definition,
  target_asfrs,
  fixed_rates = FALSE
)
}
\arguments{
\item{forecast_definition}{a pfs object}

\item{target_asfrs}{a numerical vectors of desired ASFRs at target}

\item{fixed_rates}{should the rates be fixed at jumpoff (default=FALSE)}
}
\value{
a matrix of central ASFR predictions over the forecasting horizon.
columns denote ages, rows denote forecast time steps.
}
\description{
Predict Trajectory to Target ASFRs
}
\examples{
forecast_definition <- DefineForecast(
    jumpoff_asfrs = c(0.0031, 0.0269, 0.0671, 0.0888, 0.0512, 0.0134, 0.0012),
    forecast_horizon = 30,
    ages = c(15, 20, 25, 30, 35, 40, 45),
    wlast = 5,
    target_tfr = 1.51,
    target_mab = 32.9,
    asfr_growth_rate = 0.3,
    timestep_of_steepest_growth = 15,
    randomness = 'finland1995-2024'
 )
target_asfrs <-
  PredictTargetASFRs(forecast_definition)$optimized_target_asfrs
PredictTrajectoryToTargetASFRs(forecast_definition, target_asfrs)
}
