% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preestimated_covariance}
\alias{preestimated_covariance}
\title{A List of Pre-estimated Covariance Matrices of ASFR Innovations}
\format{
An object of class \code{list} of length 1.
}
\source{
Own estimates based on Finnish ASFR time series.
}
\usage{
preestimated_covariance
}
\description{
A List of Pre-estimated Covariance Matrices of ASFR Innovations
}
\keyword{datasets}
