% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleRandomWalkASFRs.R
\name{SampleRandomWalkASFRs}
\alias{SampleRandomWalkASFRs}
\title{Sample Random Walk Realizations of ASFRs}
\usage{
SampleRandomWalkASFRs(forecast_definition, mean_asfr_trajectory_matrix, nsim)
}
\arguments{
\item{forecast_definition}{A pfs object.}

\item{mean_asfr_trajectory_matrix}{A matrix of central ASFR
forecasts as output by PredictTrajectoryToTargetASFRs().}

\item{nsim}{An integer scalar for the number of simulated ASFR paths.}
}
\value{
A 3d array of sampled ASFR paths.
}
\description{
Sample Random Walk Realizations of ASFRs
}
\examples{
forecast_definition <- DefineForecast(
    jumpoff_asfrs = c(0.0031, 0.0269, 0.0671, 0.0888, 0.0512, 0.0134, 0.0012),
    forecast_horizon = 30,
    ages = c(15, 20, 25, 30, 35, 40, 45),
    wlast = 5,
    target_tfr = 1.51,
    target_mab = 32.9,
    asfr_growth_rate = 0.3,
    timestep_of_steepest_growth = 15,
    randomness = 'finland1995-2024'
 )
target_asfrs <-
  PredictTargetASFRs(forecast_definition)$optimized_target_asfrs
trajectory <-
  PredictTrajectoryToTargetASFRs(forecast_definition, target_asfrs)
SampleRandomWalkASFRs(forecast_definition, trajectory, nsim = 100)
}
